+++ src/Client/ClientCore/Rendering/Scenes/MainMenu/MainMenuScene.cs
@@ -105,10 +105,12 @@ public class MainMenuScene : IScene
         globalLightPhiRad = 0.0f;
     }
 
-    public void PopulateWorldFragmentConstants(out Vector4 ambientLightColor, out Vector4 globalLightColor)
+    public void PopulateWorldFragmentConstants(out Vector4 ambientLightColor, out Vector4 globalLightColor,
+        out Vector2 globalLightShift)
     {
         ambientLightColor = Vector4.One;
         globalLightColor = Vector4.One;
+        globalLightShift = Vector2.Zero;
     }
 
     public void UpdateGui(RenderPlan renderPlan)